refactor gallery-carousel
